requirements.txt
# Core Dependencies
pymongo>=4.6.0
cryptography>=41.0.0
python-dotenv>=1.0.0

# Web Framework
Flask>=3.0.0
Flask-CORS>=4.0.0
Flask-Limiter>=3.5.0

# GUI Framework
PyQt6>=6.6.0
PyQt6-WebEngine>=6.6.0

# Additional Security
bcrypt>=4.1.0
pyotp>=2.9.0

# Utilities
colorama>=0.4.6
tabulate>=0.9.0


